+++ CHANGELOG.md
@@ -1,25 +1,37 @@
+## 3.0.0
+
+* Renamed `Transition` to `PageRouteTransition`.
+* Added `settings` parameter.
+* Added `PageRouteTransitionBuilder`.
+* Added `none` transition effect.
+* Removed `arguments` parameter.
+* Improved the directory structure and code.
+* Updated README
+* Updated example
+
 ## 2.0.2
 
-* Added Object to pass arguments to other pages
+* Added Object to pass arguments to other pages.
 
 
 ## 2.0.1
 
-* Update complete example
+* Updated README
+* Updated example
 
 ## 2.0.0
 
 * Stable null safety release.
-* Update README
-* Update example
+* Updated README
+* Updated example
 
 ## 1.0.2
 
-* Update README
+* Updated README
 
 ## 1.0.1
 
-* Fix typo
+* Fixed typo
 
 ## 1.0.0